Extracted the handling of input into its own class
